requirements.txt
streamlit
moviepy<2.0
librosa
mediapipe
openai-whisper
google-generativeai
python-dotenv
opencv-python
streamlit-option-menu
